python_backend/requirements.txt
google-adk
Pillow
imageio
pytesseract